# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DJ Scaffolding API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ContactInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  /** @minLength 1 */
  message: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  message: string;
  read?: boolean;
  createdAt: string;
}

export interface ContentSection {
  id: number;
  key: string;
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  updatedAt: string;
}

export interface ContentSectionInput {
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
}

export interface ServiceInput {
  name: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}