# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DJ Scaffolding API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form submissions
  - name: content
    description: CMS content management
  - name: auth
    description: Admin authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
        "400":
          description: Validation error

  /contact/submissions:
    get:
      operationId: listContactSubmissions
      tags: [contact]
      summary: List all contact submissions (admin)
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSubmission"

  /content/sections:
    get:
      operationId: listContentSections
      tags: [content]
      summary: List all editable content sections
      responses:
        "200":
          description: All content sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentSection"

  /content/sections/{key}:
    get:
      operationId: getContentSection
      tags: [content]
      summary: Get a content section by key
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentSection"
        "404":
          description: Not found
    put:
      operationId: updateContentSection
      tags: [content]
      summary: Update a content section (admin)
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentSectionInput"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentSection"

  /services:
    get:
      operationId: listServices
      tags: [content]
      summary: List all services
      responses:
        "200":
          description: All services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [content]
      summary: Create a new service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    put:
      operationId: updateService
      tags: [content]
      summary: Update a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [content]
      summary: Delete a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [auth]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminSession
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ContactInput:
      type: object
      required: [name, email, message]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        phone:
          type: ["string", "null"]
        service:
          type: ["string", "null"]
        message:
          type: string
          minLength: 1

    ContactSubmission:
      type: object
      required: [id, name, email, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        service:
          type: ["string", "null"]
        message:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time

    ContentSection:
      type: object
      required: [id, key, title, content, updatedAt]
      properties:
        id:
          type: integer
        key:
          type: string
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time

    ContentSectionInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]

    Service:
      type: object
      required: [id, name, description, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer

    ServiceInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer

    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean