# lib/api-zod/src/generated/types/contactSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DJ Scaffolding API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  service?: string | null;
  message: string;
  read?: boolean;
  createdAt: Date;
}