# lib/api-zod/src/generated/types/contentSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DJ Scaffolding API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentSection {
  id: number;
  key: string;
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  updatedAt: Date;
}