# lib/api-zod/src/generated/types/contentSectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DJ Scaffolding API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentSectionInput {
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}