# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DJ Scaffolding API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder: number;
}